Fix infinite loop in PREFER_IS_NOT
The check ran into an infinite loop when encountering a chained method
call like obj->method_1( )->method_2( ). The problem was not in the 
check itself but in the bracket determination not having a specific
return value variant for tokens like )->method_2( that contain both an
opening and a closing bracket.