Fix infinite loop in PREFER_IS_NOT (#31)
The PREFER_IS_NOT check ran into an infinite loop when it encountered a
statement like
```abap
if not obj->call_1( )->call_2( ).
```
due to the `)->call_2(` not being treated correctly by the bracket
analyzer methods. This PR fixes that.

(There seems to be an annoying issue where the cloud abapGit version
orders methods differently from the ZABAPGIT_STANDALONE version when
serializing, so that the diff here is larger than expected. I did not
manually reorder any methods)