Add abaplint as GitHub action (#34)
[abaplint](https://abaplint.org/) is an ABAP lint tool that doesn't need
an ABAP system (hence has a completely different use case to what we are
doing here) but its lints are also inspired by the style guide and
should help to adhere to the style guide without manual review.

In the current configuration I first activated all lints, then
successively deactivated those we don't need. Then I fixed all issues in
the current code so that we can actually start off having zero issues -
which is the reason the diff here is so large because I had to reformat
a lot of code, especially to adhere to [reasonable line
lengths](https://github.com/SAP/styleguides/blob/main/clean-abap/CleanABAP.md#stick-to-a-reasonable-line-length).

The `syntax_check` check is deactivated for some files because the 2302
version of the API the linter uses didn't yet contain some of the
methods/fields we're using there. We can remove this exclusion once the
2308 API is publicly available.

Also I _think_ I have configured the workflow correctly so that abaplint
will be run on each new pull request in the future but there might still
be config missing, so if the next time you open a pull request after
this is merged, please pay attention to whether there's an abaplint
check running and if not, notify me and I'll have to look at this again.

Since I just looked at the issues the linter actually produced when
setting up the config, it's possible that there are still lints active
that don't make sense for us but that didn't produce any issues on the
current code. We can just deactivate those once they occur.